crawl.py ----Assignment Context Investigation
This branch is intended for Tommy.  I added better comments in general and comments specifically for Tommy.  This is so when he begins adding the investigation part of ELSA into ELSA, he can see my thought process, what my code does, what it does poorly, etc.  

Instead of committing directly to the master branch, I decided to make this a new branch.  I'm not too sure what all we would like our Github to contain.  By creating a new branch, we can either commit these changes to the master branch (if we would like to keep copies of this type of thing) or we can just delete the branch.  More on this later.

Tommy, please keep in mind the following:
    - We are currently rebuilding build.  Because of this, the current build models will not work with the new system.  
    - If you would like to implement a quick fix solution like one of these were, please feel free so we can get this part up and working for our in-house work.  However, I would also like you to think about the changes listed in the documents.  As we rebuild build, we need to rebuild crawl, and build remove, and a bunch of other work.
    - Are there other parts of NASA that we could crawl to gather beneficial information for ELSA?  

Notice [for quick solution to problem]:  A function already exists to gather instrument_host data from an investigation product.  This could be used (in the poor manner as noted in the comments in regards to the instrument_host function) to also grab the investigation data we seek.